feat: 使用雪花ID

+++ src/main/kotlin/core/security/SensitiveRoutePlugin.kt
@@ -37,7 +37,7 @@ fun Route.sensitive(
         onCall { call ->
             val principal = call.principal<UserPrincipal>() ?: return@onCall
 
-            val user = userRepository.findById(UserId(principal.userId)).getOrNull()
+            val user = userRepository.findById(UserId(principal.userId.toLong())).getOrNull()
 
             if (user == null) {
                 logger.warn("Sensitive route: user not found, userId={}", principal.userId)
